refactor: 마이페이지에서 사용되고 있는 MyCategoryTab 컴포넌트 user밑으로 이동 시킴
