+++ frontend/src/pages/MyPage/index.tsx
@@ -1,10 +1,10 @@
 import React, { Suspense, useState } from 'react';
 
 import Loading from '@/components/@common/Loading/Loading';
+import MyCategoryTab from '@/components/user/MyCategoryTab/MyCategoryTab';
 import ArticleBox from '@/components/user/UserArticleBox/UserArticleBox';
 import UserProfile from '@/components/user/UserProfile/UserProfile';
 import useGetUserInfo from '@/hooks/user/useGetUserInfo';
-import MyCategoryTab from '@/pages/MyPage/MyCategoryTab/MyCategoryTab';
 import * as S from '@/pages/MyPage/index.styles';
 import { CategoryType } from '@/types/myPage';
